requirements.txt
arrow==0.13.1
dash==0.40.0
Flask==1.0.2
kafka-python==1.4.4
pandas==0.23.4
plotly==3.4.0
requests==2.21.0
